refactor: move classes to entities folder
